src/test/kotlin/Day19Test.kt
import org.eruanno.Day19
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.CsvSource

class Day19Test {
    private val subject = Day19()

    @ParameterizedTest
    @CsvSource(
        "'day-19-test', 6",
        "'day-19', 371"
    )
    fun testPartOne(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partOne(testFile))
    }

    @ParameterizedTest
    @CsvSource(
        "'day-19-test', '16'",
        "'day-19', '0'"
    )
    fun testPartTwo(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partTwo(testFile))
    }
}
